[RRBE-30] Recovery & Retry
>> Recover From Consumer Failure <<
If the consumer process stops processing and leaves entries in the
Pending Entry List, they are recovered and processed before new
entries are processed.

>> Retry <<
If the consumer does not read data before its blocking time expires,
we want to retry and increase the blocking time several times.  If no
data is received after a reasonable amount of time, stop the consumer,
displaying a message to the user about its abort.